Add install-time checking of platform support
Rather than blindly allowing the user to attempt an install on a
system that we know will fail with confusing template errors, we
now perform validation and issue either a user-friendly error
(if the compiler is known to lack the necessary support) or a
user-friendly warning (if we are not sure).  This checking can be
overridden in all cases with UPCXX_INSTALL_NOCHECK=1.

Output lots of relevant platform information at the top of the install
output, to help in bug report triage. This output can be suppressed
with UPCXX_INSTALL_QUIET=1.