nobs: Update default GASNet-EX download link
For the rest of this month, nobs will default to downloading
the collaborator-snapshot branch of GASNet-EX.

These stable snapshots currently carry the version identifier "2017.6.1"
and already contain many fixes relative to the June beta.

The snapshot will eventually contain EX release candidates and an
updated version number.

The URL will be updated again to the final release URL in the
last week of September.

CC: @PHHargrove