Initial upcxx-run unified launch script
Simple script that tries to find the right launch command for the
given UPC++ executable, based on code originally by Paul.

Usage:
  upcxx-run <num_ranks> <executable> [arguments...]

upcxx-run is available in the utils source directory, but only
guaranteed to work if one or more of the following holds:

1. $GASNET_PREFIX points to valid GASNet install or build directory, or
2. The script resides in a full UPC++ install, or
3. $UPCXX_INSTALL points to a full UPC++ install, or
4. The GASNet launchers are in the user's $PATH.
(and this is the current priority order)

This version is very minimal and just ensures we have something -
many enhancements should eventually be added.