Simplify the prog-guide Makefile
This commit removes some of the unnecessary generality from
example/prog-guide/Makefile.  The idea is to keep it deliberately
minimal so the Makefile itself can serve as an example of a good way to
build a simple UPC++ program.

Cleanup some misleading comments

Move use of EXTRA_FLAGS to avoid option duplication