issue 156: temporary workaround for a CI regression
upcxx-run is currently looking for an ident string which has changed
names and format in the current GEX head, breaking the nightly CI.

This workaround effectively changes the behavior where the ident is missing
(ie new GEX code) from "fatal error" to "assume the value is large enough",
since the new configure-time default size of "85% of physical memory shared
host-wide" is sufficient for all common use cases.

This change is backwards-compatible - the behavior for old GEX code (including
current collaborator-snapshot) is unchanged. Also, detection of non-GEX
executables is still provided by the conduit detection ident string
logic.

This is NOT a complete solution to issue #156, just a temporary stop-gap
measure to preserve CI testing until a proper solution can be deployed.