Resolve issue #165 (run-tests problem on Crays)
This commit resolves issue #165 (run-tests not portable to many Cray
systems) by the simple expedient of commenting-out the three tests
which are problematic (cannot safely be launched w/ current
infrastructure).

Testing on Theta at ALCF (where the original problem was observed)
shows the problem to be resolved.  The only required departure from
the documented instructions was to set the GASNET environment variable
to point to a pre-staged GASNet-EX tarball.  This was necessary
because batch and interactive jobs run without outside network access.