Strengthen the GCC pattern in system-checks
The gcc --version string is highly flexible, depending on details of
packaging and installation, so the pattern has to be wider than we'd
like in order to capture all reasonable GCC installs. In particular,
we cannot even rely upon the strings "gcc" or "g++" appearing.
However, we can presumably rely on the Free Software Foundation
copyright thanks to their militant licensing, so detect that to help
restrict the pattern.