issue 118: enforce GEX version interlock at compile time
This preprocessor code ensures that if we reach UPC++ library compile
time with an incompatible version of GASNet/GEX, we issue an
understandable error message near the top of the errors.

This is admittedly inferior to early validation in nobs that could
provide a shorter error sooner in the process, however it ensures
that we never get silent misbehavior from using an incompatible
GEX version (as currently happens if you pass GEX v2018.9 to develop).