issue #190: teach upcxx-run to honor UPCXX_PYTHON
This is a "minimal" fix to issue #190 that merely ensures upcxx-run
at least tries $UPCXX_PYTHON before dropping dead due to the `python`
interpreter in the PATH being too old.

It still requires that `python` exists in PATH (or is invoked
explicitly to process the script).

Also, improve the error message when upcxx-run cannot find a
compliant Python interpreter