Introducing CUDA support. To enable, set UPCXX_CUDA=1 in environment during nobs or install. The installer does not yet sweep cuda=0/1 as a configuration dimension, that's TODO.
test/copy.cpp is the only cuda aware test. Run with 2
ranks. If you have more than one gpu, hack the file to
set:
  constexpr int dev_n=2,3,etc;

- test/copy.cpp does validate the byte values are preserved.

- It does not report effective badnwidth, thats TODO.

This has only been tested on summit. It is very possible
that nobs will fail to build elsewhere since it parses
the output of "nvcc --verbose" on a brittle way. Portability
testing is TODO.