Fix link error on min_aligment in device_allocator::allocate in non-opt mode
This defect is due to a [change in definition behavior for static
constexpr variables](https://en.cppreference.com/w/cpp/language/static)
between C++11 to C++17. The variable in question is always odr-used, but
the reference gets optimized away in ndebug mode so never makes it to
link-time.

Also fix a std::min call that I believe should be std::max to compute an
alignment value compatible with all constraints.