examples/prog-guide: Team and atomics fixups
Adding team PG examples to the implementation repository. Fixed the
unclear progress loop in dmap-atomics example and made a similar change
in the dmap-quiescence example

Added a statement about the odd number of processors. made the snippets
smaller

Added a comment stating that the example requires an even number of
processes