Remove unneeded conduit disable configure arguments
This commit removes the explicit disable of several conduits which are
not currently supported in GASNet-EX.  This is no longer necessary
because GASNet-EX's configure script no longer fails when the
corresponding development headers and libraries are present.

Note that we expect ofi-conduit to reappear in GASNet-EX in the
future.  So, at least the `--disable-ofi` needs to be removed before
that happens.