fix issue #184: SEGV on put_flood
As previously written, the put_flood output loop was treading into UB,
using a std::initializer_list beyond its lifetime.
I've confirmed via printf that in -O3 mode, gcc-9.1.0/x86_64/Linux was
overwriting the contents of the std::initializer_list before the end
of the loop, resulting in passing of a garbage `char *` pointer to
the `how` argument of `make_row`, precipitating a later SEGV in hash.