upcxx: Rework -M handling
gcc's -M and -MM options are for dependency generation and disable
code generation and linkage, so correctly detect those.

The other gcc -M* options either require one of the two above, or
(in the case of -MD and -MMD) still allow codegen and linking.

This more targetted detection resolves the problem that the overly
general pattern was leading to incorrect behavior with PGI options
such as -Mipa -Msignextend -Mscalapack etc.