Update allocate calls to new signature
The signatures for `upcxx::allocate()` and `device_allocator<Device>::allocate()`
have changed, in accordance with
[Spec pull request #25](https://bitbucket.org/berkeleylab/upcxx-spec/pull-requests/25).

The `alignment` parameter has moved from being a final defaulted
template argument to being a final defaulted function argument.
This enables the use of alignment values which are not constexpr.

This is technically a breaking change, but should only affect programs
that explicitly set the alignment parameter or rely upon explicitly
typed expressions from pointers to these functions.