issue #278: workaround compile failures on test/regression/issue138.cpp
test/regression/issue138.cpp constructs a deep chain of template
instantiations, which generates a combinatorial explosion of code in the
object file, especially when optimizations that reduce text size are
disabled.

A new `-DMINIMAL` preprocessor option activates a version of the test
that elides some of the types tested, pruning the combinatorial space
and trading test coverage for code size (and the ability to compile at
all on affected platforms).

The only other change is a manual string combining optimization, again
to reduce object code size.