install: Fix an unreported bug with PrgEnv-gnu handling
Previously the install script was not correctly detecting PrgEnv-gnu,
due to a scripting error. As a result the intended CROSS validation
warning was skipped and CC/CXX were defaulted incorrectly.

This defect was previously masked by nobs unconditionally ignoring
the CC/CXX variables for CROSS targets, so the only potential impact
of this defect on users of the last release is a missing check to
warn PrgEnv-gnu users to set CROSS.