INSTALL: document our bash version floor
It's possible that older bashes also work, but this is the oldest
that we have in current testing.

The key point here is none of our scripts should unconditionally
rely upon the large set of features added in bash 4.x, mostly
because Apple is still shipping 3.2.

For a good summary of bash version-isms, see:
https://mywiki.wooledge.org/BashFAQ/061