Keep just a single list of tests
This modifies `run-tests` to query the lists of seq and par tests
from `Makefile.rules`, to keep a single centralized instance of
those lists.

If in the future we want differing coverage between `make check`
and `run-tests`, then `Makefile.rules` can still be the single
location for what could become 4 lists instead of 2.