Expand the library ident strings
Expand the ident strings in libupcxx.a to provide better
self-idenfication of configuration information.

UPC++ executables normally contain mostly equivalent information from
libgasnet.a, but this provides key information in libupcxx.a to assist
in debugging potential link failures caused by misconfiguration.

This commit also notably embeds the C++ compiler identity and
C++ language level in use, which are not embedded in libgasnet.a.

Example output:

```
$ ident libupcxx.a

libupcxx.a:
     $UPCXXLibraryVersion: 20190901L $
     $UPCXXNetwork: IBV $
     $UPCXXThreadMode: SEQ $
     $UPCXXCodeMode: debug $
     $UPCXXGASNetVersion: 2019.9.0 $
     $UPCXXCUDAEnabled: 0 $
     $UPCXXAssertEnabled: 1 $
     $UPCXXMPSCQueue: atomic $
     $UPCXXCompilerID: |COMPILER_FAMILY:GNU|COMPILER_VERSION:9.2.0|COMPILER_FAMILYID:1|STD:__STDC__,__cplusplus=201103L|misc:9.2.0| $
     $UPCXXCompilerStd: 201103L $
     $UPCXXBuildTimestamp: Dec 23 2019 15:39:28 $
```