Ensure correct operation of 'make -d' and 'make --debug=...'
We must strip the options out of MAKEFLAGS in the environment of a
sub-make from which we capture output (just echovar).  Note that if
MAKEFLAGS would begin with a single '-', then GNU Make omits it.  So
we must strip bare 'd' (but not 'debug' or '-debug', because that
would actually be self-consistent!).