Improve `exe` and `run` targets
If `$(SRC)` is not an absolute path and is not found as a relative
path, then we now search `test/` and `example/` in the source
directory.  The search is such that both `SRC=regression/issue219.cpp`
and `SRC=issue219.cpp` will work (and the existing logic was already
such that they hash to the same cached executable).

Also now provide a clear error if $(SRC) not found at all.