upcxx.sh: new "BUILDDIR" mode
This adds logic to util/upcxx.sh to special-case UPCXX_META=BUILDDIR
(literal).  In this case the script parses its arguments and echoes
the resulting var=val of any UPCXX_* variables.  This is for use in
(coming soon) a `$(top_blddir)/bin/upcxx` which need those variables
to ensure the correct libraries are built.  This approach ensures we
need only maintain a single copy of the arg parsing logic.