Simplify user messages regarding `make`
This commit changes the helpful "Recommended next steps" messages to
use just `make` or `gmake` in place of their full path, IFF these were
located in `$PATH`.  If the user provided a value (via environment or
configure command line), the full value of `$GMAKE` is still used.

Additionally, the search order is changed to check `make` prior to
`gmake` with the expectation that this will yield the user's most
familiar command more times than not.