Add `upcxx-meta` to `$(top_blddir)/bin`
This commit adds `$(top_blddir)/bin/upcxx-meta` to the set of
`configure` outputs.  This `upcxx-meta` is fully-featured, including
`DUMP` and `SET` support.  With this addition, the build-dir scripts
should be sufficient for even Makefile-based bug reproducers.

Since the purpose of `upcxx-meta` is to generate machine-readable
output on `stdout`, the "INFO: ..." message must be generated on
`stderr`.  For consistency, this commit modifies the `upcxx` and
`upcxx-run` generated by `configure` to also emit this message on
`stderr`.

The relevant portion of `docs/build-devel.md` has been
updated/improved.