Order debug targets ahead of opt
This commit ensures that debug variants targets are ordered ahead of
the opt ones in lists of targets.  This ordering is intended to help
developers reach library build errors more quickly by default.

The variable UPCXX_OPTDBG has been renamed UPCXX_DBGOPT in the hope
of reducing the chance of confusion.