make check/tests: reduce prerequisite builds
This commit improves the `check` and `tests` targets to build only
the GASNet and libupcxx variants they will consume.  This means one
one codemode (`debug` unless configured with `--enable-single=opt`)
and the conduits in `$(NETWORKS)` (which will be just the default
conduit for `make check`).

Additionally, this commit factors the selection of the codemode used
to build tests.