make check/tests: parallel compiles
This commit reworks the logic for the compilation of tests which is
common to `check`, `tests` and `test_install` targets to be
parallel-make friendly.

In the non-parallel case we continue through all the failures,
building all the tests, and present a summary at the end of the
compilation.  In the parallel mode, we stop at the first failure.

The bulk of this commit involves splitting the test-related targets
off into their own Makefile.tests.  This is necessary because
Makefile.rules is explicitly `.NOTPARALLEL`, to avoid some pretty
spectacular failures for things like `make all check` or `make all
install` when both targets spawn overlapping builds.

This commit does NOT introduce parallel _execution_ of tests.