configure: fetch/verify of GASNet-EX as first step
This commit moves the fetch, unpack and verification of GASNet-EX
sources to be the first step after argument processing.  This will
enable support for `--help=recursive` in a future commit.

Additionally, this commit standardizes the format of comments at the
start of each major step in `configure`.

This commit contains only code movement and addition of comment lines.
There are no changes in the logic other than order of the major steps.