Centralize defn of dependency-generation flags
This commit replaces four individual instances of the compiler flags
for generation of dependencies with use of a macro now defined in
`compiler.mak`.  Additionally, significant documentation is provided
regarding how this particular magic works.

Prior to this commit the generation of dependencies for `make exe` and
`make run` (but not libupcxx) were broken for PGI, and a lack of this
central definition contributed to that brokenness.