make exe/run: fix dependency tracking with PGI
This commit works around the difference between PGI's treatment of
`-MT 'target1 target2'` and that of all other compiler families we
currently support.

While the approach taken here for PGI is generic enough to work with
any currently supported compiler, it is not being adopted as the
default because it requires an extra `sed` invocation in each and
every dependency-generation command.