Reduce difference between compile and dep-gen
This commit ensures that dependency generation for objects in libupcxx
will include any source-file-specific compiler flags.

Currently, such flags are only defined for `runtime.cpp` in order to
define `UPCXX_GIT_VERSION` when available.  However, this mechanism
was intended to be more general, and so it is better to ensure now
that dependency generation include it, than to have silently incorrect
dependencies later.