configure: shell checks
This commit provides checks against "explicitly wrong" shells (e.g
invoking `zsh ./configure`) or a too-old version of bash.

For a non-bash shell, we make a careful effort to exec /bin/bash.
If /bin/bash is missing, the exec fails, or it is not bash, then
we give up with a helpful message:
```
ERROR: UPC++ requires /bin/bash, which must be version 3.2 or newer.
ERROR: /bin/bash is not available or does not appear to be bash.
```

If we are running bash (or the reexec described above gets us there),
but the version is too old, the following message is given:
```
ERROR: UPC++ requires /bin/bash, which must be version 3.2 or newer.
ERROR: This bash reports a version of '3.0.2'.
```

For `tcsh ./configure` we cannot be too smart, but the failure is
at least enough to provide a bit of a clue as to the problem:
```
BASH_VERSION: Undefined variable.
```

Note that this logic is intentionally NOT trying to find bash in
`$PATH`, since at this point in time we have multiple scripts that
begin `#!/bin/bash` and no logic for automatically rewriting them.