make: check target now honors NETWORKS
The `make check` rule now honors the setting of `$(NETWORKS)` which
continues to default to the default network.  The message that
precedes the test compilation is updated not to say "default network"
unless that is actually the case.

`INSTALL.md` has also been updated to relocate  the description of
`NETWORKS` into the portion of the text describing `make check`.