upcxx-meta: rename GASNET_INSTALL to GASNET_PREFIX
This commit makes `GASNET_PREFIX` the "primary" name for the
`upcxx-meta` variable giving the location of the GASNet-EX tree, but
preserves `GASNET_INSTALL` as an alias.  The use of `GASNET_BUILD` as
a same-meaning-but-different-name variable in the build tree is
eliminated (in favor of also using `GASNET_PREFIX`).

Additionally, this commit repairs the logic which (due to entanglement
with the abandoned install-single support) was incorrectly using the
(yet to be) installed GASNet prefix in the in-build-tree `upcxx-meta`.
This went unnoticed because the were then no current consumers for the
`GASNET_BUILD` or `GASNET_INSTALL` (and none now for `GASNET_PREFIX`).