Fix install-time header crawl for canonical paths
This commit modifies the header-crawl used at install time to allow
for compilers which use canonical paths in generated dependency files.
This differs from the path we previously expected because of a symlink
between builddir and srcdir which inserts the `upcxx/` path component.
Of course, we now look for *either* the original or canonical path in
the dependency file when identifying headers to install.

This fixes an observed failure to install headers with (at least)
gcc-7.3.0 on Cygwin.