fixup: Improved configure arg parsing
An earlier commit to improve argument parsing was intended to diagnose
usage errors such as `--with-cc` and `--with-cc=` where a required
argument is missing.  However, the configure script would continue
normally after the diagnostic was issued.  This commit corrects the
behavior for a missing required argument to be fatal as intended.