CMake: Improve use of LINK_OPTIONS
Ensure that potentially proprietary compiler options are only
added to the link step when invoking the CXX compiler frontend.
Normal clients of UPC++ should always be linking using the C++ frontend,
(or a wrapper around it) because libupcxx has a hard dependence on
the libstdc++ normally added by that wrapper.

Hybrid clients (HipMer) benefit from this change which ensures we
don't try to pass C++ compiler options to the `upcc` linker wrapper
(which eventually invokes upcxx to re-add these options anyhow).