Generate upcxx_config.hpp
This commit adds logic to the build infrastructure to generate a
`upcxx_config.hpp` header and to install it.

There is a distinct header for each instance of `libupcxx.a`.

The generation takes place between compilation of  GASNet and the
build of libupcxx.a.  There is currently no mechanism to cause the
generated header to be automatically rebuilt if something changes in
GASNet.  However, changes to the config-probe script will trigger
re-generation.

This commit includes documentation on how to provide scripts for the
generation of header content.  However, no such scripts are included
(yet) and the resulting `upcxx_config.hpp` is thus empty except for
the presence of multiple-inclusion guards.

Also, there are not yet any *consumers* of the new header.