Support bash in locations other than /bin
This commit makes all the necessary changes to allow compilation of
UPC++ on a FreeBSD system with `bash` located `/usr/local/bin/bash`.
This includes having `configure` search `$PATH` for a shell to run
itself with, substitution into generated in-build-tree scripts,
setting of `SHELL` in makefiles, and rewriting of the she-bang when
installing `upcxx.sh`.

Resolves issue #295