Fix hidden_am_concurrency_level.sh probe
This commit rewrites the logic in `hidden_am_concurrency_level.sh` for
capturing a value in C code for use in shell.  The previous logic was
flawed in that it did not allow for the possibility that the
preprocessor introduces whitespace while tokenizing.  Rather than fix
this on the shell side by allowing whitespace in the regex, the new
approach uses token concatenation which cannot introduce whitespace.

There are now comments to ease reuse of the logic in other probes.

Additionally, the probe is also made more general to allow for values
of `GASNET_HIDDEN_AM_CONCURRENCY_LEVEL` greater than 1.