Re-implement atomic_domain interface to GASNet
* atomic_domain<T> now supports T = any 32/64-bit integer type,
  no longer restricted to `std::(u)int{32,64}_t`. This means it's
  now portable and safe to use things like `atomic_domain<long>`
  or `atomic_domain<size_t>`. Fixes issue #298.
* Leverage the dependency-free <gasnet_fwd.h> header (introduced in
  2018.6.0) to enable direct use of GEX_OP_*, GEX_DT_*, GEX_FLAG_* and
  gex_Event_t in the public header, without pulling in all of GASNet.
* Remove a redundant gex_Event_Test() operation

Overall this change includes removal of a table lookup and several
branches that were previously in the critical path of AMO injection
(for normal opt-mode builds not using LTO).
This is expected to provide measureable improvement to the overhead
of AMO injection using the PSHM bypass path (local_team peers).