configure and upcxx: improved --version output
This commit updates both `configure` and `upcxx` to print the version
in the form `YYYY.M.P` (with no leading zeros on M or P) rather than
as `YYYYMMPPL` (with leading zeros and a literal `L`).

Additionally, the case of `upcxx --version` in the build tree now
operates correctly, where it would previously print an empty version.

Finally, the behavior has been made less sensitive to whitespace and
will make a reasonable effort to print _something_ if there is a
define of `UPCXX_VERSION` not matching the expected pattern.