Colorization of compiler output
This commit implements a `COLOR=1` option to several make targets to
request colorization of compiler output, even when that output is not
to a (p)tty.  This is useful, for instance, when piping to a pager.

1. Adds a configure probe to find the compiler flags required to force
colorization of output.  Does not assume `CC` and `CXX` are identical.
There is currently only one "spelling", but other possibilities are
simple to add.

2. Adds makefile logic to honor `make COLOR=1 ...` set on the command
line (intentionally _ignoring_ the same in the environment) to enable
passing of the probed compiler flags in recipes.  Without this setting
the probed flags are discarded and the changes below pass nothing
additional (relative to before this commit) except for whitespace.

3. Adds colorization flags to `MANUAL_{C,CXX}FLAGS` for the GASNet
builds.

4. Adds colorization flags to `LIBUPCXX_{C,CXX}_COMPILE` used for
compiling objects in `libupcxx`.

5. Adds colorization flags to the command lines used by the `exe` and
`run` targets.  These are *not* included in `EXTRAFLAGS`, ensuring
that executables built with and without colorized output are the same
cache entry.

6. Adds colorization flags to the commands used in the `check`,
`tests` and `test_install` targets.