Correct `make -j check` for newer GMake
This commit ensures `make check` will actually run in parallel on all
supported versions of GNU Make.

Prior to this commit, GNU Make 4.2 would build tests serially due to a
change to `$(MAKEFLAGS)`.  Where the presence of any `-j` or `-jN` was
previously advertised in `$(MAKEFLAGS)` as simply `-j`, the latest
versions preserve the `N` when given.  Use of a wildcard `%` ensures
both will match.