Remove the global_ptr hack from upcxx::copy(T*)
Initial testing with the new global_ptr checks immediately uncovered a
deliberate breakage of the global_ptr invariants in the implementation of
upcxx::copy (which was appropriately described as a "Hack!" in the comments).

Replace this hack with a version that passes the decomposed gptr fields to a
detail function instead of forging a broken object.  IMHO the result is
cleaner, and the behavior should be equivalent - possibly even a tiny bit more
efficient by eliding the construction of the invalid global_ptr's.

Add gptr checks to the input arguments to upcxx::copy