issue #173: upcxx-run executable search and error messages
This commit alters the handling of the GEX executable identified by
upcxx-run (the first command line argument that leads to a valid file
containing the ident strings).  Currently PATH is NOT searched when
identifying the executable -  this is intentional but could be added
as an extension later if we wanted.

If the executable is a bare filename in the current directory, then
`./` is prepended to that argument, ensuring it will be found when
we later pass it to `exec` or an underlying spawner.

Fixes issue #173