Repair GASNet-EX version validation
This commit resolves an issue discovered late in pre-release testing
in which the new configure script could not validate the GASNet-EX API
version in (at least) GASNet-EX 2018.12.0.  Since we still claim to
support GASNet-EX that old, we are breaking code freeze to make this
small surgical change to avoid needlessly raising the floor on
GASNet-EX version support.

This change has been validated against the tarballs for the following
GASNet-EX tarballs:

+  2018.6.0 and 2018.9.0 fail gracefully
+  2018.12.0, 2019.3.0, 2019.6.0, 2019.9.0 and Stable are OK