Generalize and simplify GNU version check
This commit generalizes the logic used to probe the GNU toolchain
version used (by non-GNU compilers in particular) to support probes of
arbitrary C or C++ compilers (but mainly `CC` and `CXX`).

The values of `CFLAGS` or `CXXFLAGS` are now included in the probe
commands.

Additionally, the preprocessor "goop" has been slightly simplified.