configure: useful defaults on Cray front-ends
This commit modifies the default CC/CXX on a Cray front-end (as
determined by the presence of a loaded supported PrgEnv module) to
behave as a standard Linux system if CROSS is unset.  Previously
we were using the Cray wrappers, resulting in a build that was
not appropriate to either the front-end or the compute nodes.

The warning about setting CROSS to build for the compute-nodes
is NOT changed.