configure: improve the warning for dubious dirs
This commit augments the warning issued when interpreters like
`python` and `bash` (to be used in `#!` in installed scripts) are
located in directories outside a whitelist.

The main change is to describe the alternative of using the "env trick"
to defer the lookup in `$PATH`.

Additionally the handling of arguments to interprets (such as like
`bash --norc`) is corrected.  Where Linux seems fine with `A='/a/b c';
dirname $A`, macOS issues the `dirname` usage message.  So, we now
trim off the args.