configure: recommend --with-cross=, not CROSS env var
This commit updates the warning issued when `configure` is run on
a Cray front-end without a cross-configure setting, to recommend the
new command line syntax rather than the old (still supported)
environment variable.

Additionally, the warning now lists both supported values, to lessen
the chance of a user blindly copying the SLURM value on an ALPS-based
system.