docs: revise Intel/libstdc++ recommendation
This commit revises the Intel section of `docs/alt-compilers.md` to
advise use of *both* `-gxx-name` and `-gcc-name` options.  This is
consistent with the new error and warning messages given by the
`configure` script due to the behavior seen on (but not known to be
exclusive to) Cray systems.

(cherry picked from commit f2b83601087708e1cd14a8ffc2cba3d1706ca9db)