system-checks.sh: wrap long messages
This commit breaks especially long arguments to `echo` (for warnings
and error messages) into multiple argument on multiple lines.  This is
intended to improve readably.

This commit is entirely cosmetic, no changes have been made to logic
or to the content of the wrapped messages.  In particular, this does
NOT apply wrapping to the message the end-user will see.