configure: automatically cross-compile on Cray XC
This commit resolves issue 353 by adding logic to intelligently
default `$CROSS`.  This includes "auto completion" of `cray-aries`, as
well as fully-automatic cross-compilation if the environment "looks
right" and no `--with-cross=...` option was specified..

A new option `--without-cross` is available in case one wants to build
for the front-end of a Cray XC, or for another "ancillary" system such
as Cori's GPU nodes.