configure: allow (unsupported) use of PrgEnv-pgi
This commit makes our handling of PrgEnv-pgi on Cray XC consistent
with that of other compilers where we lack sufficient positive or
negative information to confidently identify them as COMPILER_GOOD or
COMPILER_BAD.  Previously we would fail at configure-time if
cross-compiling for the Cray XC with PrgEnv-pgi loaded.  Now we emit
the warning containing the supported compilers and allow the user to
continue at their own peril.

There are no associated documentation changes, as we were already
silent on the status of PrgEnv-pgi.

Limited testing of PrgEnv-pgi on Piz-Daint shows no new issues specific
to that platform, meaning COMPILER_BAD is not appropriate. However, we
lack the resources to perform periodic runs (though compile-only tests
are possible) necessary to make a confident claim of COMPILER_GOOD (a
prerequisite for labelling this as a supported platform).

This partially addresses issue 352.