upcxx-run: relaxed validation for udp custom spawn
This commit makes the validation of `GASNET_SSH_SERVERS` for
udp-conduit applications conditional on `GASNET_SPAWNFN` being either
unset or `S`.  In particular, no validation is done for the case of
`GASNET_SPAWNFN=C`.

Previously, if one set `GASNET_SPAWNFN=C` to use a custom spawner,
then the validation of `GASNET_SSH_SERVERS` was performed even though
the value was (probably) not used.  This was not a problem if the
value was empty, or set to a list with at least as many elements as
processes to launch.  However, a list with fewer elements (as would
be accepted by `ibv` and eventually `ofi` and `ucx`) would yield a
fatal error:

    Error: For UDP conduit, need to specify at least one host per
    process in -ssh-servers or environment variable
    GASNET_SSH_SERVERS, or use -localhost parameter