Rename user-facing UPCXX_CODEMODE=O3 to opt
This change is more reflective of the reality, which is to select
the optimized version of the library, which may or may not be using
exactly optimization level 3 (the value is platform-dependent).

The actual optimization level used for application code defaults
to the same used by the library, but is directly under user control
on the compiler command line.

Also, modify `upcxx` behavior so that bare `-O` or `-g` options
select the opt/debug versions of the library but don't add that
literal option to the compile line, as for some compilers that
might inadvertantly override the default -O2 or -O3 with -O1.
Users can and should explicitly specify optimization level for
their own code, passing options like `-O3 -Ofast` etc.