Makefile: fix GASNET_CONDUIT set in environment
This commit fixes a misbehavior in several developer-facing
make targets in which `GASNET_CONDUIT` set in the environment
was ignored in favor of the default (though settings on the
command line did work).

The solution is to use a new `UPCXX_DEFAULT_NETWORK` variable to
encode the default network in the generated top-level Makefile, rather
than `GASNET_CONDUIT`, thus removing all issues of precedence.