configure: update mechanism for conduit validation
This commit replaces use in `configure` of a shell `case` statement
for validation of `--with-default-network=...` with regex-based
matching against simple lists.

One benefit of this change is to simplify the addition of a conduit
(or promotion of one out of UNSUPPORTED status) by ensuring the two
place that require update (the other being a makefile fragment) are
essentially identical (ignoring different quoting and whitespace).
However, the main motivation of this change is to enable (in a
forthcoming commit) validation of `UPCXX_NETWORK` by the in-builddir
`upcxx` and `upcxx-meta`.

Additionally, this commit relocates the "not an officially supported
network" warning to the block of warnings emitted at the end of the
`configure` output (where it is far less likely to be missed).