Improved network validation by in-buildddir utils
This commit improves the error messages from the in-builddir instances
of `upcxx` and `upcxx-meta` when given an invalid or unsupported
network.  The message now provides the *configured* set of networks if
GASNet's configure has already run.

If GASNet's configure has not yet run, the output is still a list of
all supported networks, but with the addition of the qualification "a
system-specific subset of".