makefile: implement dev-{tests,check} targets
This commit adds `dev-tests` and `dev-check` make targets for use by
the developer.  These new targets differ from the un-suffixed versions
in two key ways.  The first is that the list of tests is a superset.
The second is that testing iterates over all four pairs of codemode
and threadmode (excluding some tests as appropriate).

New `dev-run-tests` and `dev-tests-clean` are also provided.