makefile: correct `make check UPCXX_CODEMODE=opt`
This commit corrects a misbehavior in the `check` target when
explicitly requesting the "opt" codemode prior to building the
corresponding "debug" instance(s) of `libupcxx.a`.

The symptom was a fail to run due to attempted use of a `upcxx-run` in
a debug directory which has not been built previously and *will not*
be build dynamically due to `$(UPCXX_SKIP_REBUILD)`.

The fix is simply to ensure the in-builddir `upcxx-run` will honor the
`UPCXX_CODEMODE` when set.

The use of `UPCXX_CODEMODE=opt` with `make check` does *not* appear in
user documentation (only in `build-devel.md`).