Removed some unnecessary moves of user types from the rpc path.
Removed an unnecessary copy of user type on rpc path for the case when the
rpc lambda returns a non-future (by switching to future1::then_lazy)

Added test "test/rpc-ctor-trace.cpp" to count number of various ctors
that a type experiences dynamically along an rpc roundtrip. This test
asserts that the number of default and copy constructions meets the
expected value.