CI scripts for use by developers
This commit adds a set of scripts to exercise `make dev-tests` and
`make dev-run-tests` on a variety of platforms, accessible by the
UPC++ developers, which cover the range of supported compiler families
and their respective supported ranges.  These include the appropriate
per-platform configure and run scripts.

These scripts are intended to simplify testing by developers prior to
(or during) the PR processes.