legacy install: serialize GASNet-EX configures
This commit modifies the legacy `install` script to perform the two
GASNet-EX configure steps sequentially by default.  This is in
recognition that (1) a significant fraction of user-error cases (bad
inputs) are likely to result in failures in one or both of these
configure steps and (2) that running both configures in parallel makes
it very difficult for the user to parse the output (and thus to
diagnose/resolve their own problems).